Rename regex parser lib
