+++ src/analysers/tools/regex-parser.ts
